requirements.txt
pandas~=0.24.2
torchvision~=0.15.1+cu117
torch~=2.0.0+cu117
matplotlib~=3.7.1
numpy~=1.24.1
pillow~=9.3.0
scipy~=1.10.1
scikit-learn~=1.3.0